changing labscript types from long to int; reading long values from h5 files are problematic

+++ modules/hdf5-wrapper/include/shot-file.h
@@ -13,7 +13,7 @@
 #include <unordered_map>
 #include <vector>
 
-using LabscriptDictValueTypes = boost::variant<std::string, long, bool, double>;
+using LabscriptDictValueTypes = boost::variant<std::string, int, bool, double>;
 using LabscriptDictType =
     std::unordered_map<std::string, LabscriptDictValueTypes>;
